third_party/blink/web_tests/platform/win/external/wpt/css/css-text/i18n/zh/css-text-line-break-zh-pr-normal-expected.txt

This is a testharness.js-based test.
[FAIL] 00B1  PLUS-MINUS SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 92 +/- 1 but got 62
[FAIL] 20AC  EURO SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 92 +/- 1 but got 62
[FAIL] 2116  NUMERO SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 100 +/- 1 but got 70
[FAIL] FE69  SMALL DOLLAR SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FF04  FULLWIDTH DOLLAR SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE1  FULLWIDTH POUND SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE5  FULLWIDTH YEN SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE6  FULLWIDTH WON SIGN may NOT appear at line start if zh and normal
  assert_approx_equals: expected 99 +/- 1 but got 69
Harness: the test ran to completion.
